append info to separate tables

+++ public/scripts/client.js
@@ -4,7 +4,7 @@ function onReady() {
     getList();
     $('#submitButton').on('click', addTask);
     $('#container').on('click', '.completeButton', completeTask);
-    $('#container').on('click', '.deleteButton', deleteTask);
+    $('#container, #container2').on('click', '.deleteButton', deleteTask);
 }
 
 function getList() {
@@ -38,27 +38,31 @@ function addTask() {
 
 function appendTasks(taskArray) {
     $('#container').empty();
+    $('#container2').empty();
     for (var i in taskArray) {
         var $row = $('#container').append('<tr></tr>');
-        if (taskArray[i].completed) {                   // set class to completed to unfinished
-            $row.append('<td class="completed">' + taskArray[i].task + '</td>');
-        } else {
+        var $row2 = $('#container2').append('<tr></tr>');
+        if (taskArray[i].completed) {                   // append completed tasks
+            $row2.append('<td class="completed">' + taskArray[i].task + '</td>');
+            $row2.append('<td><button class="deleteButton" data-id="' + taskArray[i].id + '">Delete</button></td>');
+            $('#container2').append($row2);
+        } else {                                    // append unfinished tasks
             $row.append('<td class="unfinished">' + taskArray[i].task + '</td>');
+            $row.append('<td><button class="completeButton" data-id="' + taskArray[i].id + '"data-completed=' + taskArray[i].completed + '>Completed</button></td>');
+            $row.append('<td><button class="deleteButton" data-id="' + taskArray[i].id + '">Delete</button></td>');
+            $('#container').append($row);
         }
-        $row.append('<td><button class="completeButton" data-id="' + taskArray[i].id + '"data-completed=' + taskArray[i].completed + '>Complete</button></td>');
-        $row.append('<td><button class="deleteButton" data-id="' + taskArray[i].id + '">Delete</button></td>');
-        $('#container').append($row);
     }
 }   // end appendTasks funct
 
 function completeTask() {
-    var taskId = $(this).data('id');          
+    var taskId = $(this).data('id');
     var completed = $(this).data('completed');      // yields 'true' or 'false' string
     if (completed) {
         console.log('already completed!');      //  append 'already completed!' to DOM !!
         return;
     } else {
-        console.log('false... need to run a PUT call!');        
+        console.log('false... need to run a PUT call!');
         // ajax PUT call
         $.ajax({
             method: 'PUT',
@@ -73,14 +77,19 @@ function completeTask() {
 } // end completeTask funct
 
 function deleteTask() {              // send id of task to be deleted to server
-    var taskId = $(this).data('id');     // grab id of deleted button/row, store as INT
-    $.ajax({
-        method: 'DELETE',
-        url: '/tasks/' + taskId,
-        data: taskId,
-        success: function (res) {
-            console.log(res);
-            getList();
-        }
-    });
+    var confirmed = confirm('Are you sure you want to delete this item?');  //returns boolean
+    if (confirmed) {
+        var taskId = $(this).data('id');     // grab id of deleted button/row, store as INT
+        $.ajax({
+            method: 'DELETE',
+            url: '/tasks/' + taskId,
+            data: taskId,
+            success: function (res) {
+                console.log(res);
+                getList();
+            }
+        });
+    } else {
+        return;
+    }
 } // end deleteTask funct
